Change which event blocks were created in
