+++ src/main/java/gregtech/common/blocks/MetaBlocks.java
@@ -25,6 +25,7 @@ import gregtech.common.blocks.foam.BlockPetrifiedFoam;
 import gregtech.common.blocks.wood.BlockGregPlanks;
 import gregtech.common.blocks.wood.BlockRubberLeaves;
 import gregtech.common.blocks.wood.BlockRubberLog;
+import gregtech.common.blocks.wood.BlockGregPlanks;
 import gregtech.common.blocks.wood.BlockRubberSapling;
 import gregtech.common.pipelike.cable.BlockCable;
 import gregtech.common.pipelike.cable.Insulation;
@@ -219,50 +220,6 @@ public class MetaBlocks {
         PLANKS = new BlockGregPlanks();
         PLANKS.setRegistryName("planks");
 
-        StoneType.init();
-
-        createGeneratedBlock(m -> m.hasProperty(PropertyKey.DUST) && m.hasFlag(GENERATE_FRAME), MetaBlocks::createFrameBlock);
-        createGeneratedBlock(m -> m.hasProperty(PropertyKey.ORE) && m.hasProperty(PropertyKey.DUST), MetaBlocks::createSurfaceRockBlock);
-
-        createGeneratedBlock(
-                material -> (material.hasProperty(PropertyKey.INGOT) || material.hasProperty(PropertyKey.GEM))
-                        && !OrePrefix.block.isIgnored(material),
-                MetaBlocks::createCompressedBlock);
-
-        for (Material material : GregTechAPI.MATERIAL_REGISTRY) {
-
-            if (material.hasProperty(PropertyKey.ORE)) {
-                createOreBlock(material);
-            }
-
-            if (material.hasProperty(PropertyKey.WIRE)) {
-                for (BlockCable cable : CABLES) {
-                    if (!cable.getItemPipeType(null).isCable() || !material.getProperty(PropertyKey.WIRE).isSuperconductor())
-                        cable.addCableMaterial(material, material.getProperty(PropertyKey.WIRE));
-                }
-            }
-            if (material.hasProperty(PropertyKey.FLUID_PIPE)) {
-                for (BlockFluidPipe pipe : FLUID_PIPES) {
-                    if (!pipe.getItemPipeType(pipe.getItem(material)).getOrePrefix().isIgnored(material)) {
-                        pipe.addPipeMaterial(material, material.getProperty(PropertyKey.FLUID_PIPE));
-                    }
-                }
-            }
-            if (material.hasProperty(PropertyKey.ITEM_PIPE)) {
-                for (BlockItemPipe pipe : ITEM_PIPES) {
-                    if (!pipe.getItemPipeType(pipe.getItem(material)).getOrePrefix().isIgnored(material)) {
-                        pipe.addPipeMaterial(material, material.getProperty(PropertyKey.ITEM_PIPE));
-                    }
-                }
-            }
-        }
-        for (BlockFluidPipe pipe : FLUID_PIPES) {
-            if (!pipe.getItemPipeType(pipe.getItem(Materials.Wood)).getOrePrefix().isIgnored(Materials.Wood) ||
-                    !pipe.getItemPipeType(pipe.getItem(Materials.TreatedWood)).getOrePrefix().isIgnored(Materials.TreatedWood)) {
-                pipe.addPipeMaterial(Materials.Wood, new FluidPipeProperties(310, 5, false));
-                pipe.addPipeMaterial(Materials.TreatedWood, new FluidPipeProperties(310, 8, false));
-            }
-        }
         registerTileEntity();
 
         //not sure if that's a good place for that, but i don't want to make a dedicated method for that
@@ -270,90 +227,6 @@ public class MetaBlocks {
         Blocks.FIRE.setFireInfo(RUBBER_LOG, 5, 5);
         Blocks.FIRE.setFireInfo(RUBBER_LEAVES, 30, 60);
         Blocks.FIRE.setFireInfo(PLANKS, 5, 20);
-    }
-
-    /**
-     * Deterministically populates a category of MetaBlocks based on the unique registry ID of each qualifying Material.
-     *
-     * @param materialPredicate a filter for determining if a Material qualifies for generation in the category.
-     * @param blockGenerator    a function which accepts a Materials set to pack into a MetaBlock, and the ordinal this
-     *                          MetaBlock should have within its category.
-     */
-    protected static void createGeneratedBlock(Predicate<Material> materialPredicate,
-                                               BiConsumer<Material[], Integer> blockGenerator) {
-
-        Map<Integer, Material[]> blocksToGenerate = new TreeMap<>();
-
-        for (Material material : GregTechAPI.MATERIAL_REGISTRY) {
-            if (materialPredicate.test(material)) {
-                int id = material.getId();
-                int metaBlockID = id / 16;
-                int subBlockID = id % 16;
-
-                if (!blocksToGenerate.containsKey(metaBlockID)) {
-                    Material[] materials = new Material[16];
-                    Arrays.fill(materials, Materials.NULL);
-                    blocksToGenerate.put(metaBlockID, materials);
-                }
-
-                blocksToGenerate.get(metaBlockID)[subBlockID] = material;
-            }
-        }
-
-        blocksToGenerate.forEach((key, value) -> blockGenerator.accept(value, key));
-    }
-
-    private static void createCompressedBlock(Material[] materials, int index) {
-        BlockCompressed block = new BlockCompressed(materials);
-        block.setRegistryName("meta_block_compressed_" + index);
-        for (Material material : materials) {
-            COMPRESSED.put(material, block);
-        }
-    }
-
-    private static void createFrameBlock(Material[] materials, int index) {
-        BlockFrame block = new BlockFrame(materials);
-        block.setRegistryName("meta_block_frame_" + index);
-        for (Material m : materials) {
-            FRAMES.put(m, block);
-        }
-    }
-
-    private static void createSurfaceRockBlock(Material[] materials, int index) {
-        BlockSurfaceRock block = new BlockSurfaceRock(materials);
-        block.setRegistryName("meta_block_surface_rock_" + index);
-        for (Material material : materials) {
-            SURFACE_ROCK.put(material, block);
-        }
-    }
-
-    private static void createOreBlock(Material material) {
-        StoneType[] stoneTypeBuffer = new StoneType[16];
-        int generationIndex = 0;
-        for (StoneType stoneType : StoneType.STONE_TYPE_REGISTRY) {
-            int id = StoneType.STONE_TYPE_REGISTRY.getIDForObject(stoneType), index = id / 16;
-            if (index > generationIndex) {
-                createOreBlock(material, copyNotNull(stoneTypeBuffer), generationIndex);
-                Arrays.fill(stoneTypeBuffer, null);
-            }
-            stoneTypeBuffer[id % 16] = stoneType;
-            generationIndex = index;
-        }
-        createOreBlock(material, copyNotNull(stoneTypeBuffer), generationIndex);
-    }
-
-    private static <T> T[] copyNotNull(T[] src) {
-        int nullIndex = ArrayUtils.indexOf(src, null);
-        return Arrays.copyOfRange(src, 0, nullIndex == -1 ? src.length : nullIndex);
-    }
-
-    private static void createOreBlock(Material material, StoneType[] stoneTypes, int index) {
-        BlockOre block = new BlockOre(material, stoneTypes);
-        block.setRegistryName("ore_" + material + "_" + index);
-        for (StoneType stoneType : stoneTypes) {
-            GregTechAPI.oreBlockTable.computeIfAbsent(material, m -> new HashMap<>()).put(stoneType, block);
-        }
-        ORES.add(block);
     }
 
     public static void registerTileEntity() {
